# swagger.json
{
    "openapi":  "3.0.0",
    "info":  {
                 "title":  "RevUp Job Portal API",
                 "version":  "3.0.0",
                 "description":  "## RevUp — AI-Powered Job Portal API",
                 "contact":  {
                                 "name":  "RevUp Dev Team"
                             }
             },
    "servers":  [
                    {
                        "url":  "http://localhost:5000",
                        "description":  "Local Development Server"
                    }
                ],
    "components":  {
                       "securitySchemes":  {
                                               "bearerAuth":  {
                                                                  "type":  "http",
                                                                  "scheme":  "bearer",
                                                                  "bearerFormat":  "JWT",
                                                                  "description":  "Enter your JWT token. Obtain it from POST /api/auth/login"
                                                              }
                                           },
...
                   },
    "security":  [
                     {
                         "bearerAuth":  [

                                        ]
                     }
                 ],
...
    "paths":  {
...
                  "/api/admin/jobs":  {
                                          "get":  {
                                                      "tags":  [
                                                                   "Admin"
                                                               ],
                                                      "summary":  "Get all jobs (open and closed)",
                                                      "security":  [
                                                                       {
                                                                           "bearerAuth":  [

                                                                                          ]
                                                                       }
                                                                   ],
                                                      "parameters":  [
                                                                         {
                                                                             "in":  "query",
                                                                             "name":  "status",
                                                                             "schema":  {
                                                                                            "type":  "string",
                                                                                            "enum":  [
                                                                                                         "open",
                                                                                                         "closed"
                                                                                                     ]
                                                                                        }
                                                                         }
                                                                     ],
                                                      "responses":  {
                                                                        "200":  {
                                                                                    "description":  "List of all jobs"
                                                                                }
                                                                    }
                                                  }
                                      },
                  "/api/admin/stats/jobs":  {
                                                "get":  {
                                                            "tags":  [
                                                                         "Admin"
                                                                     ],
                                                            "summary":  "Detailed job analytics",
                                                            "security":  [
                                                                             {
                                                                                 "bearerAuth":  [

                                                                                                ]
                                                                             }
                                                                         ],
                                                            "responses":  {
                                                                              "200":  {
                                                                                          "description":  "Job analytics"
                                                                                      }
                                                                          }
                                                        }
                                            },
...
                  "/api/admin/jobs/{id}":  {
                                               "delete":  {
                                                              "tags":  [
                                                                           "Admin"
                                                                       ],
                                                              "summary":  "Force-delete any job post",
                                                              "security":  [
                                                                               {
                                                                                   "bearerAuth":  [

                                                                                                  ]
                                                                               }
                                                                           ],
                                                              "parameters":  [
                                                                                 {
                                                                                     "in":  "path",
                                                                                     "name":  "id",
                                                                                     "required":  true,
                                                                                     "schema":  {
                                                                                                    "type":  "integer"
                                                                                                }
                                                                                 }
                                                                             ],
                                                              "responses":  {
                                                                                "200":  {
                                                                                            "description":  "Job deleted"
                                                                                        }
                                                                            }
                                                          }
                                           },
...
                  "/api/applications/job/{jobId}":  {
                                                        "get":  {
                                                                    "tags":  [
                                                                                 "Applications"
                                                                             ],
                                                                    "summary":  "Get all applicants for a specific job (Recruiter only)",
                                                                    "security":  [
                                                                                     {
                                                                                         "bearerAuth":  [

                                                                                                        ]
                                                                                     }
                                                                                 ],
                                                                    "parameters":  [
                                                                                       {
                                                                                           "in":  "path",
                                                                                           "name":  "jobId",
                                                                                           "required":  true,
                                                                                           "schema":  {
                                                                                                          "type":  "integer"
                                                                                                      }
                                                                                       }
                                                                                   ],
                                                                    "responses":  {
                                                                                      "200":  {
                                                                                                  "description":  "Applicants list"
                                                                                              },
                                                                                      "403":  {
                                                                                                  "description":  "Forbidden - Not your job"
                                                                                              }
                                                                                  }
                                                                }
                                                    },
...
                  "/api/jobs/recommended":  {
                                                "get":  {
                                                            "tags":  [
                                                                         "Jobs"
                                                                     ],
                                                            "summary":  "Get skill-matched jobs ranked by percentage (Seeker only)",
                                                            "security":  [
                                                                             {
                                                                                 "bearerAuth":  [

                                                                                                ]
                                                                             }
                                                                         ],
                                                            "responses":  {
                                                                              "200":  {
                                                                                          "description":  "Recommended jobs with match_percentage"
                                                                                      }
                                                                          }
                                                        }
                                            },
                  "/api/jobs/my-postings":  {
                                                "get":  {
                                                            "tags":  [
                                                                         "Jobs"
                                                                     ],
                                                            "summary":  "All jobs posted by the logged-in recruiter",
                                                            "security":  [
                                                                             {
                                                                                 "bearerAuth":  [

                                                                                                ]
                                                                             }
                                                                         ],
                                                            "responses":  {
                                                                              "200":  {
                                                                                          "description":  "Recruiter job list"
                                                                                      }
                                                                          }
                                                        }
                                            },
                  "/api/jobs":  {
                                    "get":  {
                                                "tags":  [
                                                             "Jobs"
                                                         ],
                                                "summary":  "List open jobs with search, filters, and pagination",
                                                "parameters":  [
                                                                   {
                                                                       "in":  "query",
                                                                       "name":  "search",
                                                                       "schema":  {
                                                                                      "type":  "string"
                                                                                  }
                                                                   },
                                                                   {
                                                                       "in":  "query",
                                                                       "name":  "job_type",
                                                                       "schema":  {
                                                                                      "type":  "string",
                                                                                      "enum":  [
                                                                                                   "Full-time",
                                                                                                   "Part-time",
                                                                                                   "Contract",
                                                                                                   "Internship",
                                                                                                   "Remote",
                                                                                                   "Hybrid"
                                                                                               ]
                                                                                  }
                                                                   },
                                                                   {
                                                                       "in":  "query",
                                                                       "name":  "location",
                                                                       "schema":  {
                                                                                      "type":  "string"
                                                                                  }
                                                                   },
                                                                   {
                                                                       "in":  "query",
                                                                       "name":  "page",
                                                                       "schema":  {
                                                                                      "type":  "integer",
                                                                                      "default":  1
                                                                                  }
                                                                   },
                                                                   {
                                                                       "in":  "query",
                                                                       "name":  "limit",
                                                                       "schema":  {
                                                                                      "type":  "integer",
                                                                                      "default":  10
                                                                                  }
                                                                   }
                                                               ],
                                                "responses":  {
                                                                  "200":  {
                                                                              "description":  "Job list with pagination"
                                                                          }
                                                              }
                                            },
                                    "post":  {
                                                 "tags":  [
                                                              "Jobs"
                                                          ],
                                                 "summary":  "Post a new job (Recruiter only)",
                                                 "security":  [
                                                                  {
                                                                      "bearerAuth":  [

                                                                                     ]
                                                                  }
                                                              ],
                                                 "requestBody":  {
                                                                     "required":  true,
                                                                     "content":  {
                                                                                     "application/json":  {
                                                                                                              "schema":  {
                                                                                                                             "type":  "object",
                                                                                                                             "required":  [
                                                                                                                                              "title",
                                                                                                                                              "description"
                                                                                                                                          ],
                                                                                                                             "properties":  {
                                                                                                                                                "title":  {
                                                                                                                                                              "type":  "string"
                                                                                                                                                          },
                                                                                                                                                "description":  {
                                                                                                                                                                    "type":  "string"
                                                                                                                                                                },
                                                                                                                                                "location":  {
                                                                                                                                                                 "type":  "string"
                                                                                                                                                             },
                                                                                                                                                "job_type":  {
                                                                                                                                                                 "type":  "string"
                                                                                                                                                             },
                                                                                                                                                "salary_range":  {
                                                                                                                                                                     "type":  "string"
                                                                                                                                                                 },
                                                                                                                                                "skillIds":  {
                                                                                                                                                                 "type":  "array",
                                                                                                                                                                 "items":  {
                                                                                                                                                                               "type":  "integer"
                                                                                                                                                                           }
                                                                                                                                                             }
                                                                                                                                            }
                                                                                                                         }
                                                                                                          }
                                                                                 }
                                                                 },
                                                 "responses":  {
                                                                   "201":  {
                                                                               "description":  "Job created"
                                                                           }
                                                               }
                                             }
                                },
...
                  "/api/jobs/{id}/status":  {
                                                "patch":  {
                                                              "tags":  [
                                                                           "Jobs"
                                                                       ],
                                                              "summary":  "Toggle job status open/closed (Recruiter only)",
                                                              "security":  [
                                                                               {
                                                                                   "bearerAuth":  [

                                                                                                  ]
                                                                               }
                                                                           ],
                                                              "parameters":  [
                                                                                 {
                                                                                     "in":  "path",
                                                                                     "name":  "id",
                                                                                     "required":  true,
                                                                                     "schema":  {
                                                                                                    "type":  "integer"
                                                                                                }
                                                                                 }
                                                                             ],
                                                              "responses":  {
                                                                                "200":  {
                                                                                            "description":  "Status toggled"
                                                                                        }
                                                                            }
                                                          }
                                            },
...
              }
}